Agent/Smart_Reserach_Assistant_Project/requirements.txt
fastapi
uvicorn
langgraph
google-generativeai
pydantic
markdown
textblob
python-dotenv
gradio
httpx
